+++ PiPPl/Utility/LocalVideoLibraryManager.swift
@@ -10,32 +10,6 @@ import Photos
 import UIKit
 import Dispatch
 
-class ThumbnailMemoryCache {
-
-    static let shared = ThumbnailMemoryCache()
-
-    private init() {}
-
-    private var thumbnailCache = NSCache<NSString, UIImage>()
-
-    func thumbnail(for asset: PHAsset) -> UIImage? {
-        return thumbnailCache.object(forKey: asset.localIdentifier as NSString)
-    }
-
-    func setThumbnail(_ image: UIImage, for asset: PHAsset) {
-        thumbnailCache.setObject(image, forKey: asset.localIdentifier as NSString)
-    }
-
-    func removeThumbnail(for asset: PHAsset) {
-        thumbnailCache.removeObject(forKey: asset.localIdentifier as NSString)
-    }
-
-    func removeAllThumbnails() {
-        thumbnailCache.removeAllObjects()
-    }
-
-}
-
 class ThumbnailDiskCache {
 
     static let shared = ThumbnailDiskCache()
